refactor: move DeleteGroupDrawer up to the higher layer
